requirements.txt
asgiref==3.8.1
Django==4.2.17
django-ckeditor-5==0.2.15
django-js-asset==2.2.0
pillow==11.0.0
sqlparse==0.5.1
tzdata==2024.2
